Refactor partition formulation
