+++ src/omlt/neuralnet/__init__.py
@@ -1,5 +1,5 @@
 from omlt.neuralnet.network_definition import NetworkDefinition
 from omlt.neuralnet.nn_formulation import (FullSpaceNNFormulation, ReducedSpaceNNFormulation,
                                            FullSpaceSmoothNNFormulation, ReducedSpaceSmoothNNFormulation,
-                                           ReluBigMFormulation, ReluComplementarityFormulation)
-                                           #ReluPartitionFormulation)
+                                           ReluBigMFormulation, ReluComplementarityFormulation,
+                                           ReluPartitionFormulation)
